r/aws_neptune_cluster_endpoint: Move functions around.
