--- internal/service/neptune/id.go
@@ -1,17 +1,0 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: MPL-2.0
-
-package neptune
-
-import (
-	"fmt"
-	"strings"
-)
-
-func readClusterEndpointID(id string) (clusterIdentifier string, endpointIndetifer string, err error) {
-	idParts := strings.Split(id, ":")
-	if len(idParts) != 2 {
-		return "", "", fmt.Errorf("expected ID in format clusterIdentifier:endpointIndetifer, received: %s", id)
-	}
-	return idParts[0], idParts[1], nil
-}
